backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.9.1
openai>=2.15.0
langsmith==0.7.3
sse-starlette==2.1.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
httpx==0.27.2
python-multipart>=0.0.6

# Document parsing libraries (Module 6: Multi-Format Support)
pypdf==4.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
html2text==2024.2.26
